refactor vec reflection and turned floats to double

+++ includes/color.h
@@ -5,14 +5,14 @@
 
 typedef struct s_color
 {
-	float	r;
-	float	g;
-	float	b;
+	double	r;
+	double	g;
+	double	b;
 } t_color;
 
 
 void color_clamp(t_color *color);
-void color_mul_scalar(t_color *color, float scalar);
+void color_mul_scalar(t_color *color, double scalar);
 void color_add(t_color *color1, t_color *color2);
 void color_mul(t_color *color1, t_color *color2);
