+++ parser/parser_utils.c
@@ -73,7 +73,7 @@ bool parse_vec3(t_vec3 *vec, char *str)
 	return (true);
 }
 
-static bool parse_color_channel(float *value, char *str)
+static bool parse_color_channel(double *value, char *str)
 {
 	int result;
